refactor(serve): decompose connection.ts handlers into domain-aligned modules (swamp-club#893)
Split the 66 request handlers out of src/serve/connection.ts (5,975
lines) into seven domain modules under src/serve/handlers/ — data,
model, workflow, vault, access, report, admin — plus handlers/shared.ts
for the cross-cutting utilities (ConnectionContext, send, sendError,
sanitizeErrorForClient, authorizeOrReject, isAccessModelType, shared
constants). connection.ts (now 1,621 lines) remains the dispatcher:
Zod schemas, validateServerRequest, handleConnection, the
handleMessage switch, and back-compat re-exports.

Pure move — no behavior change:
- All 71 moved functions byte-identical to the originals (only an
  `export` keyword added); dispatcher and all 68 schema blocks
  byte-identical.
- Conservation checks exact: authorizeOrReject 69/69, sendError
  170/170, handlers 66/66.
- Shared utilities live in handlers/shared.ts rather than
  connection.ts (deviation from the issue text) to keep imports
  one-way; the public surface is preserved via re-exports, so
  mod.ts, serve.ts, and connection_test.ts are untouched.
