+++ src/serve/handlers/data_handlers.ts
@@ -1,0 +1,591 @@
+// Swamp, an Automation Framework
+// Copyright (C) 2026 Elder Swamp Club, Inc.
+//
+// This file is part of Swamp.
+//
+// Swamp is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License version 3
+// as published by the Free Software Foundation, with the Swamp
+// Extension and Definition Exception (found in the "COPYING-EXCEPTION"
+// file).
+//
+// Swamp is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Swamp.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * Data-domain request handlers (data.* and summarise verbs).
+ */
+
+import {
+  consumeStream,
+  createDataDeleteDeps,
+  createDataGetDeps,
+  createDataListDeps,
+  createDataRenameDeps,
+  createDataVersionsDeps,
+  createLibSwampContext,
+  createSummariseDeps,
+  dataDelete,
+  dataGet,
+  dataList,
+  dataQuery,
+  type DataQueryDeps,
+  dataRename,
+  dataSearch,
+  type DataSearchDeps,
+  dataVersions,
+  parseDuration,
+  summarise,
+} from "../../libswamp/mod.ts";
+import type {
+  DataDeletePayload,
+  DataGetPayload,
+  DataListPayload,
+  DataQueryPayload,
+  DataRenamePayload,
+  DataSearchPayload,
+  DataVersionsPayload,
+  SummarisePayload,
+} from "../protocol.ts";
+import { findDefinitionByIdOrName } from "../../domain/models/model_lookup.ts";
+import { createDefinitionId } from "../../domain/definitions/definition.ts";
+import { ModelType } from "../../domain/models/model_type.ts";
+import type { Principal } from "../../domain/access/principal.ts";
+import {
+  authorizeOrReject,
+  type ConnectionContext,
+  DEFAULT_QUERY_LIMIT,
+  MAX_QUERY_RESULTS,
+  sanitizeErrorForClient,
+  send,
+  sendError,
+} from "./shared.ts";
+
+export async function handleDataGet(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataGetPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  const resourceName = payload.modelIdOrName ?? "*";
+  const dataFields: Record<string, unknown> = {};
+  if (payload.modelIdOrName) dataFields.name = payload.modelIdOrName;
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: resourceName,
+      fields: dataFields,
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createDataGetDeps(
+      ctx.repoDir,
+      undefined,
+      ctx.repoContext.unifiedDataRepo,
+      ctx.repoContext.workflowRepo,
+    );
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataGet(libCtx, deps, {
+        modelIdOrName: payload.modelIdOrName,
+        dataName: payload.dataName,
+        workflowName: payload.workflowName,
+        runId: payload.runId,
+        version: payload.version,
+        includeContent: payload.includeContent ?? true,
+        repoDir: ctx.repoDir,
+      }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Data not found");
+      return;
+    }
+
+    send(socket, {
+      type: "data.get",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_get_failed", message);
+  }
+}
+
+export async function handleDataQuery(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataQueryPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "*",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const queryService = ctx.repoContext.dataQueryService;
+    const deps: DataQueryDeps = {
+      query: (pred, opts) => queryService.query(pred, opts),
+    };
+
+    const limit = Math.min(
+      payload.limit ?? DEFAULT_QUERY_LIMIT,
+      MAX_QUERY_RESULTS,
+    );
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataQuery(libCtx, deps, {
+        predicate: payload.predicate,
+        select: payload.select,
+        limit,
+      }),
+      {
+        resolving: () => {},
+        match: () => {},
+        projected_match: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "data.query",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_query_failed", message);
+  }
+}
+
+export async function handleDataList(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataListPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  const resourceName = payload.modelIdOrName ?? "*";
+  const listFields: Record<string, unknown> = {};
+  if (payload.modelIdOrName) listFields.name = payload.modelIdOrName;
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: resourceName,
+      fields: listFields,
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createDataListDeps(
+      ctx.repoDir,
+      undefined,
+      ctx.repoContext.unifiedDataRepo,
+      undefined,
+      ctx.repoContext.workflowRepo,
+    );
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataList(libCtx, deps, {
+        modelIdOrName: payload.modelIdOrName,
+        workflowName: payload.workflowName,
+        runId: payload.runId,
+        typeFilter: payload.typeFilter,
+      }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "No data found");
+      return;
+    }
+
+    send(socket, {
+      type: "data.list",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_list_failed", message);
+  }
+}
+
+export async function handleDataSearch(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  controller: AbortController,
+  principal: Principal | null,
+  payload?: DataSearchPayload,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "*",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const definitionRepo = ctx.repoContext.definitionRepo;
+    const dataRepo = ctx.repoContext.unifiedDataRepo;
+
+    const deps: DataSearchDeps = {
+      findAllGlobal: () => dataRepo.findAllGlobal(),
+      findDefinitionById: (type, defId) =>
+        definitionRepo.findById(
+          ModelType.create(type.normalized),
+          createDefinitionId(defId),
+        ),
+      findDefinitionByIdOrName: (idOrName) =>
+        findDefinitionByIdOrName(definitionRepo, idOrName),
+    };
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataSearch(libCtx, deps, {
+        query: payload?.query,
+        type: payload?.type,
+        lifetime: payload?.lifetime,
+        ownerType: payload?.ownerType,
+        workflow: payload?.workflow,
+        model: payload?.model,
+        contentType: payload?.contentType,
+        since: payload?.since,
+        output: payload?.output,
+        run: payload?.run,
+        streaming: payload?.streaming,
+        tags: payload?.tags,
+        limit: payload?.limit ?? 50,
+      }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "data.search",
+      id: requestId,
+      payload: { data: result ?? { items: [], totalCount: 0 } },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_search_failed", message);
+  }
+}
+
+export async function handleDataVersions(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataVersionsPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  const resourceName = payload.modelIdOrName;
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: resourceName,
+      fields: { name: resourceName },
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createDataVersionsDeps(
+      ctx.repoDir,
+      undefined,
+      ctx.repoContext.unifiedDataRepo,
+    );
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataVersions(libCtx, deps, {
+        modelIdOrName: payload.modelIdOrName,
+        dataName: payload.dataName,
+      }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Data not found");
+      return;
+    }
+
+    send(socket, {
+      type: "data.versions",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_versions_failed", message);
+  }
+}
+
+export async function handleDataDelete(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataDeletePayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "write", {
+      kind: "data",
+      name: payload.modelIdOrName,
+      fields: { name: payload.modelIdOrName },
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createDataDeleteDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataDelete(libCtx, deps, {
+        modelIdOrName: payload.modelIdOrName,
+        dataName: payload.dataName,
+        version: payload.version,
+      }),
+      {
+        deleting: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Data not found");
+      return;
+    }
+
+    send(socket, {
+      type: "data.delete",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_delete_failed", message);
+  }
+}
+
+export async function handleDataRename(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: DataRenamePayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "write", {
+      kind: "data",
+      name: payload.modelIdOrName,
+      fields: { name: payload.modelIdOrName },
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createDataRenameDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      dataRename(libCtx, deps, {
+        modelIdOrName: payload.modelIdOrName,
+        oldName: payload.oldName,
+        newName: payload.newName,
+      }),
+      {
+        renaming: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "rename_failed", "Rename operation failed");
+      return;
+    }
+
+    send(socket, {
+      type: "data.rename",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "data_rename_failed", message);
+  }
+}
+
+export async function handleSummarise(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  controller: AbortController,
+  principal: Principal | null,
+  payload?: SummarisePayload,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "model",
+      name: "*",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createSummariseDeps({
+      outputRepo: ctx.repoContext.outputRepo,
+      workflowRunRepo: ctx.repoContext.workflowRunRepo,
+      dataRepo: ctx.repoContext.unifiedDataRepo,
+      definitionRepo: ctx.repoContext.definitionRepo,
+      workflowRepo: ctx.repoContext.workflowRepo,
+    });
+
+    const sinceLabel = payload?.since ?? "7d";
+    const durationMs = parseDuration(sinceLabel);
+    const cutoffDate = new Date(Date.now() - durationMs);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      summarise(libCtx, deps, {
+        since: cutoffDate,
+        sinceLabel,
+        limit: payload?.limit,
+      }),
+      {
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "summarise",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "summarise_failed", message);
+  }
+}
