+++ src/serve/handlers/vault_handlers.ts
@@ -1,0 +1,640 @@
+// Swamp, an Automation Framework
+// Copyright (C) 2026 Elder Swamp Club, Inc.
+//
+// This file is part of Swamp.
+//
+// Swamp is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License version 3
+// as published by the Free Software Foundation, with the Swamp
+// Extension and Definition Exception (found in the "COPYING-EXCEPTION"
+// file).
+//
+// Swamp is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Swamp.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * Vault-domain request handlers (vault.* verbs).
+ */
+
+import {
+  consumeStream,
+  createLibSwampContext,
+  createVaultAnnotateDeps,
+  createVaultDeleteDeps,
+  createVaultDescribeDeps,
+  createVaultGetDeps,
+  createVaultInspectDeps,
+  createVaultListKeysDeps,
+  createVaultPutDeps,
+  vaultAnnotate,
+  vaultDelete,
+  vaultDeletePreview,
+  vaultDescribe,
+  vaultGet,
+  vaultInspect,
+  vaultListKeys,
+  vaultPut,
+  vaultPutPreview,
+  vaultSearch,
+  type VaultSearchDeps,
+} from "../../libswamp/mod.ts";
+import type {
+  VaultAnnotatePayload,
+  VaultDeletePayload,
+  VaultDescribePayload,
+  VaultGetPayload,
+  VaultInspectPayload,
+  VaultListKeysPayload,
+  VaultPutPayload,
+  VaultSearchPayload,
+} from "../protocol.ts";
+import { acquireVaultSync } from "../../cli/repo_context.ts";
+import type { Principal } from "../../domain/access/principal.ts";
+import {
+  authorizeOrReject,
+  type ConnectionContext,
+  sanitizeErrorForClient,
+  send,
+  sendError,
+} from "./shared.ts";
+
+export async function handleVaultGet(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultGetPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultGetDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultGet(libCtx, deps, payload.vaultNameOrId, payload.vaultType),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Vault not found");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.get",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_get_failed", message);
+  }
+}
+
+export async function handleVaultPut(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultPutPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (payload.refreshFrom !== undefined || payload.clearRefresh) {
+    if (
+      !authorizeOrReject(socket, requestId, principal, "admin", {
+        kind: "data",
+        name: "vault",
+        fields: {},
+      }, ctx)
+    ) return;
+  } else {
+    if (
+      !authorizeOrReject(socket, requestId, principal, "write", {
+        kind: "data",
+        name: "vault",
+        fields: {},
+      }, ctx)
+    ) return;
+  }
+
+  let flush: (() => Promise<void>) | undefined;
+  try {
+    ({ flush } = await acquireVaultSync(
+      ctx.datastoreConfig,
+      ctx.syncService,
+      ctx.repoDir,
+    ));
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_put_failed", message);
+    return;
+  }
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultPutDeps(ctx.repoDir, ctx.repoContext.eventBus);
+
+    const preview = await vaultPutPreview(
+      libCtx,
+      deps,
+      payload.vaultName,
+      payload.key,
+    );
+
+    if (preview.secretExists && !payload.force) {
+      sendError(
+        socket,
+        requestId,
+        "secret_exists",
+        `Secret '${payload.key}' already exists in vault '${payload.vaultName}'. Use --force (-f) to overwrite.`,
+      );
+      return;
+    }
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultPut(libCtx, deps, {
+        vaultName: payload.vaultName,
+        key: payload.key,
+        value: payload.value,
+        overwritten: preview.secretExists,
+        refreshFrom: payload.refreshFrom,
+        refreshTtlMs: payload.refreshTtlMs,
+        clearRefresh: payload.clearRefresh,
+      }),
+      {
+        storing: () => {},
+        warning: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.put",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    if (error instanceof DOMException && error.name === "AbortError") {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+    } else {
+      const message = sanitizeErrorForClient(error);
+      sendError(socket, requestId, "vault_put_failed", message);
+    }
+  } finally {
+    if (flush) await flush();
+  }
+}
+
+export async function handleVaultDelete(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultDeletePayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "write", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  let flush: (() => Promise<void>) | undefined;
+  try {
+    ({ flush } = await acquireVaultSync(
+      ctx.datastoreConfig,
+      ctx.syncService,
+      ctx.repoDir,
+    ));
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_delete_failed", message);
+    return;
+  }
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultDeleteDeps(ctx.repoDir, ctx.repoContext.eventBus);
+
+    const preview = await vaultDeletePreview(
+      libCtx,
+      deps,
+      payload.vaultName,
+      payload.key,
+    );
+
+    if (!preview.supportsDelete) {
+      sendError(
+        socket,
+        requestId,
+        "unsupported",
+        `Vault '${payload.vaultName}' (type: ${preview.vaultType}) does not support deleting secrets`,
+      );
+      return;
+    }
+
+    if (!preview.secretExists && !payload.force) {
+      sendError(
+        socket,
+        requestId,
+        "not_found",
+        `Secret '${payload.key}' not found in vault '${payload.vaultName}'`,
+      );
+      return;
+    }
+
+    if (!preview.secretExists && payload.force) {
+      send(socket, {
+        type: "vault.delete",
+        id: requestId,
+        payload: {
+          data: {
+            vaultName: payload.vaultName,
+            secretKey: payload.key,
+            vaultType: preview.vaultType,
+            noOp: true,
+            timestamp: new Date().toISOString(),
+          },
+        },
+      });
+      return;
+    }
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultDelete(libCtx, deps, {
+        vaultName: payload.vaultName,
+        key: payload.key,
+      }),
+      {
+        deleting: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.delete",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    if (error instanceof DOMException && error.name === "AbortError") {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+    } else {
+      const message = sanitizeErrorForClient(error);
+      sendError(socket, requestId, "vault_delete_failed", message);
+    }
+  } finally {
+    if (flush) await flush();
+  }
+}
+
+export async function handleVaultDescribe(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultDescribePayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultDescribeDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultDescribe(libCtx, deps, payload.vaultNameOrId, payload.vaultType),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Vault not found");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.describe",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_describe_failed", message);
+  }
+}
+
+export async function handleVaultInspect(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultInspectPayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultInspectDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultInspect(libCtx, deps, payload.vaultName, payload.key),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(socket, requestId, "not_found", "Secret not found");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.inspect",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_inspect_failed", message);
+  }
+}
+
+export async function handleVaultListKeys(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  controller: AbortController,
+  principal: Principal | null,
+  payload?: VaultListKeysPayload,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = await createVaultListKeysDeps(ctx.repoDir);
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultListKeys(libCtx, deps, {
+        vaultName: payload?.vaultName ?? "",
+      }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.list-keys",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_list_keys_failed", message);
+  }
+}
+
+export async function handleVaultSearch(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  controller: AbortController,
+  principal: Principal | null,
+  payload?: VaultSearchPayload,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "read", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps: VaultSearchDeps = {
+      findAllVaults: () => ctx.repoContext.vaultConfigRepo.findAll(),
+    };
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultSearch(libCtx, deps, { query: payload?.query }),
+      {
+        resolving: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    send(socket, {
+      type: "vault.search",
+      id: requestId,
+      payload: { data: result ?? {} },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_search_failed", message);
+  }
+}
+
+export async function handleVaultAnnotate(
+  socket: WebSocket,
+  ctx: ConnectionContext,
+  requestId: string,
+  payload: VaultAnnotatePayload,
+  controller: AbortController,
+  principal: Principal | null,
+): Promise<void> {
+  if (
+    !authorizeOrReject(socket, requestId, principal, "write", {
+      kind: "data",
+      name: "vault",
+      fields: {},
+    }, ctx)
+  ) return;
+
+  try {
+    const libCtx = createLibSwampContext();
+    const deps = createVaultAnnotateDeps(
+      ctx.repoDir,
+      ctx.repoContext.eventBus,
+    );
+
+    // Convert labels from string[] to Record<string,string> if provided
+    const labelsRecord: Record<string, string> | undefined = payload.labels
+      ? Object.fromEntries(payload.labels.map((l) => [l, ""]))
+      : undefined;
+
+    let result: Record<string, unknown> | undefined;
+    await consumeStream(
+      vaultAnnotate(libCtx, deps, {
+        vaultName: payload.vaultName,
+        key: payload.key,
+        url: payload.url,
+        notes: payload.notes,
+        labels: labelsRecord,
+        removeLabels: payload.removeLabels,
+        clear: payload.clear ?? false,
+      }),
+      {
+        annotating: () => {},
+        completed: (e) => {
+          result = e.data as unknown as Record<string, unknown>;
+        },
+        error: (e) => {
+          throw new Error(e.error.message);
+        },
+      },
+    );
+
+    if (controller.signal.aborted) {
+      sendError(socket, requestId, "cancelled", "Operation was cancelled");
+      return;
+    }
+
+    if (!result) {
+      sendError(
+        socket,
+        requestId,
+        "vault_annotate_failed",
+        "Vault annotation failed",
+      );
+      return;
+    }
+
+    send(socket, {
+      type: "vault.annotate",
+      id: requestId,
+      payload: { data: result },
+    });
+  } catch (error) {
+    const message = sanitizeErrorForClient(error);
+    sendError(socket, requestId, "vault_annotate_failed", message);
+  }
+}
